develop-dc2-11.0.0--->master

See merge request cn/vmoment/vmoment-meeting!3

+++ meeting-api/src/main/java/com/tiens/api/vo/VMMeetingCredentialVO.java
@@ -19,7 +19,7 @@ public class VMMeetingCredentialVO implements Serializable {
     /**
      * 过期时间戳
      */
-    private Long expireTime;
+    private Integer expireTime;
     /**
      * 随机字符串
      */
